+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Update
 - Code Cleanup
 - Temperature Service for Battery percentage to create HomeKit automations. (HomeKit doesn't allow for automations off of battery percentage)
+- Bumped dependencies.
 
 ## 1.1.2
 Update